requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.3.0
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
